Created subdirectories for past and upcoming events, changed appropriate import statements
