+++ src/components/ApplicationViews.js
@@ -1,16 +1,18 @@
 import React from "react";
 import { Route } from "react-router-dom";
-// import { UpcomingEventList } from "./events/UpcomingEventList";
-import { NewEventForm } from "./events/NewEventForm";
+//
 import { EventProvider } from "./events/EventProvider";
 import { UserProvider } from "./users/UserProvider";
 import { UserEventsProvider } from "./userEvents/UserEventsProvider";
-import { PastEventList } from "./events/PastEventList";
-import { EditPastEventForm } from "./events/EditPastEventForm";
-import { EditUpcomingEventForm } from "./events/EditUpcomingEventForm";
-import { CompactUpcomingEventList } from "./events/compactUpcomingEventList";
-// import { PastEventDetail } from "./events/PastEventDetail";
-
+//
+import { UpcomingEventList } from "./events/upcoming/UpcomingEventList";
+import { EditUpcomingEventForm } from "./events/upcoming/EditUpcomingEventForm";
+//
+import { NewEventForm } from "./events/NewEventForm";
+//
+import { PastEventList } from "./events/past/PastEventList";
+import { EditPastEventForm } from "./events/past/EditPastEventForm";
+//
 export const ApplicationViews = () => {
   return (
     <>
@@ -19,8 +21,7 @@ export const ApplicationViews = () => {
           <UserEventsProvider>
             {/* HOME / UPCOMING */}
             <Route exact path="/upcoming">
-              {/* <UpcomingEventList /> */}
-              <CompactUpcomingEventList />
+              <UpcomingEventList />
             </Route>
 
             <Route exact path="/upcoming/edit/:eventId(\d+)">
